add more user commands
some _extremely_ messy code here to try more structures out for sending display commands from the arduino to the led modules. My current goal is a 'splash' pattern out from the point of contact- the cell being touched brightens until it is close to full brightness, at which point the surrounding ones start, repeat.

I think to do this in a way that looks nice I won't be able to sequentially tell each cell to start fading on, since there may be upwards of 10 outer cells trying to turn on at the same time. Instead I'll need to start telling the next ring to prepare to turn on while fading on the prior ring, and use a global message to start them all.

Other thoughts:
perhaps 'q' command should come with an ID, and then '@' would also come with an ID, rather than a number of commands to run. Or both an ID and number of commands. So then I could queue up multiple 'zones' simultaneously.

splash intersections could be very tricky to handle well. I'll need to think about that- might need to keep a copy of the status of every cell on the main board to keep track of what should happen to each on every 'frame'.

I wonder if it may be fast enough to send a global message containing the information for every cell on a schedule, 60 times a second or something. I'll need to consider how much data each cell needs to keep things running smoothly...