change version to a 'should_save' variable to make writes to eeprom more intentional
also did some timing study to clock in (this current board) to a steady 60 fps animation timing. Some investigation will need to be done with other boards to see how consistent that is- maybe we'll need to calibrate every animation loop for each board separately if we want them to stay reasonably in sync, maybe not.

with 'synced rainbow' mode, the config data is being sent many times per second. If this was written to eeprom every time we would wear it down quite quickly.

sending the entire config object just to update the hue is quite inefficient, but since we're animating at 60fps we can afford to be slow. Currently we take 200ms to send the config object every 600ms. If we need to increase the syncing frequency to deal with out of sync clock speeds, we may need to add ways to not send the entire config object. Man, it is SO simple though